+++ web_travel/templates/country/add.html
@@ -1,4 +1,4 @@
-{% include 'base.html'%}
+{% extends 'base.html'%}
 
 {% block content %}
     {% with messages = get_flashed_messages() %}
@@ -12,7 +12,7 @@
     {% endwith %}
     <div class="row">
         <div class="col-8">
-            <form action="{{ url_for('admin.adding_country') }}" method='POST'>
+            <form action="{{ url_for('country.adding_process') }}" method='POST'>
                 {{ form.hidden_tag() }}
                 <div class="form-group">
                     {{ form.country_name.label }}
